Removed all references to BezierPath.
Included a new Jython version because 2.5b3 could not 
access Path getters.

+++ src/nodebox/handle/PointHandle.java
@@ -28,8 +28,8 @@ public class PointHandle extends AbstractHandle {
     }
 
     public void draw(GraphicsContext ctx) {
-        double x = node.asFloat(xName);
-        double y = node.asFloat(yName);
+        float x = node.asFloat(xName);
+        float y = node.asFloat(yName);
         drawDot(ctx, x, y);
     }
